Implement a device-host messaging service (#11)
Add a basic building block that can be used to let layers act as a network client to a server running on the host.

The design for this module exposes the server as a number of message-oriented endpoint services which are multiplexed on a single socket communications channel. Support message types are synchronous send (tx), asynchronous send (tx_async), and synchronous send-receive (tx_rx).

On Android the device-side socket can use an abstract Unix domain sockets, which can connect to a TCP/IP socket on the host when using "adb reverse" to proxy the connection. On Linux the device-side socket can directly use TCP/IP.

The default host server, which provides some basic services out-of-the-box, is implemented in Python. Client-side implementations of higher-level services, such as a virtual file system, will be provided in follow-on PRs.