Intercept device functions via vkGetInstanceProcAddr() (#22)
Vulkan allows device functions to be (less efficiently) intercepted via
vkGetInstanceProcAddr() instead of vkGetDeviceProcAddr(). This PR adds
support for intercepting these correctly.

This includes a custom vkGetDeviceImageMemoryRequirementsKHR() to skip
driver invocations if the driver dispatch pointer is nullptr. This is a
workaround for a bug in Unreal Engine which fetches this function using
vkGetInstanceProcAddr() with a Vulkan 1.1 instance (when it is available
and is therefore non-null) and then tries to use it later with a device
created from a Vulkan 1.0 instance (when it is not available because
VK_KHR_maintenance4 is a Vulkan 1.1 extension).