Refactor Python utilities (#37)
This PR merges an upgrade of the Python utilities with a much more capable
set of common helper functions, and a new common installer in the root
directory which can be used to install multiple layers at the same time. 

This is primarily useful to allow Khronos validation to be installed alongside
our layers, but can be used to install multiple layers too.