Workaround pre-instance functions on Android (#53)
Pre-instance functions are not supposed to be intercepted, but Android
doesn't implement the better interface for chainable pre-instance
functions. Trying to use vkGetInstanceProcAddr to get the "next layer 
down" vkEnumerateInstanceExtensionProperties so a layer can query
what extensions are available will not work for any layer other than the
bottom layer in the stack.

If you have multiple layers in installed, the higher layers will use
vkGetInstanceProcAddr to get the function in the next layer down and
get the stub implementation for the extensions the layer itself 
implements without the request reaching the driver.