Split LCSRenderPassContinuation from LCSRenderPass.
No need to carry around the extra state uncessarily.

This also makes clear from the type, rather than from the magic
tagID value that the renderpass is a continuation. This makes
refactoring to separate out the the workload as a data object
from operations on that data safer since they will not
need to propogate this magic knowledge forward.