Use VUL safe_struct (#118)
The Vulkan API accepts a lot of const pointers to input structure trees, many of which have unstructured content due to the use of opaque pNext pointers to support extensions. Layers often want to change input data before calling down the stack, and the current code just uses const_cast on the inputs and modifies in place. This is obviously violating the API contract with the application so we need to fix it.

This PR will clone structure trees where needed, using the safe_struct library from the Vulkan Utility Libraries project.

Fixes #56