Add blocking wait to comms creation (#141)
The current code has a racy initialization of the VkDevice communications module, and can crash if a second device sees a valid commsModule but gets a nullptr for commsWrapper which is still being built by the first context.

This change adds locking around comms creation to ensure that any device sees a completely inited comms stack. Lock is always taken for sake of simplicity - this isn't a performance critical code path.