Direct driver calling for no-op intercepts (#145)
Prior to this PR, layers would always intercept all known Vulkan functions, routing through a dummy pass-through implementation in the common code if per-layer code did not provide a more specialized implementation. This has a runtime cost in terms of additional dispatch indirection.

With this PR, layers now only intercept functions that are required to implement the layer. These may be mandatory functions in the common code, or per-layer functions with a <user_tag> specialization. This dispatch optimization is optional and, although on by default, can still be disabled to support exploratory API tracing for tech support investigations.

Fixes #14.