Use safe-struct chain info copy (#158)
This fixes use of multiple layers, which were failing to create an 
instance if more than one layer was in use.

When we refactored the code to use safe-struct to avoid const-casting 
in layer code we patched the incorrect chain info, modifying the 
original user chain info rather than modifying the the safe-struct clone. 

This happens to work for a single layer because the driver doesn't use
the chain, but it fails as soon as you have multiple layers because the 
chain is never advanced and getProcAddr points at the wrong layer.